Modifica el tipo de datos que retornan las funciones de los luchadores y agrega unit test basicos de cada clase.

+++ src/main/java/isaac/rodriguez/Subzero.java
@@ -6,7 +6,7 @@ public class Subzero extends Fighter{
         super(name);
     }
 
-    public float congelamiento(){
+    public int congelamiento(){
         return 40;
     }
